Rearrange jclapp
